extracted orderDetails as a presentation component displaying order properties
